Build fix.
Canonical link: https://commits.webkit.org/13852@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@16502 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/mac/FontMac.mm
@@ -483,7 +483,7 @@ void Font::drawComplexText(GraphicsContext* graphicsContext, const TextRun& run,
     // ATSUI can't draw beyond -32768 to +32767 so we translate the CTM and tell ATSUI to draw at (0, 0).
     CGContextRef context = graphicsContext->platformContext();
 
-    float colors[4];
+    CGFloat colors[4];
     graphicsContext->pen().color().getRGBA(colors[0], colors[1], colors[2], colors[3]);
     static CGColorSpaceRef rgbColorSpace = CGColorSpaceCreateDeviceRGB();
     
@@ -598,7 +598,7 @@ void Font::drawGlyphs(GraphicsContext* context, const FontData* font, const Glyp
     wkSetCGFontRenderingMode(cgContext, drawFont);
     CGContextSetFontSize(cgContext, 1.0f);
 
-    float colors[4];
+    CGFloat colors[4];
     context->pen().color().getRGBA(colors[0], colors[1], colors[2], colors[3]);
     static CGColorSpaceRef rgbColorSpace = CGColorSpaceCreateDeviceRGB();
